+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed running GUI with args
 - Fixed default dock/tab settings
 
+### Removed
+
+- Removed "Unpack mods" setting
+
 ## [0.7.0]
 
 ### Added